Board Component
Created the board component. Generates a random word from an array along with the appropriate amount of spaces (called in Square) by breaking up the word into an array of chars. Still working on the filter function (displaying the guessed consonant or vowel in the appropriate place) need placeholder buttons. Added to the main GUI also (Game.js).